Fixed datatype for PriceTotal in OrderInfo
Switched from INTEGER to DECIMAL

+++ schemagen.sql
@@ -30,7 +30,7 @@ CREATE TABLE OrderInfo (
     OrderID INTEGER NOT NULL,
     ItemID INTEGER NOT NULL,
     ItemAmount INTEGER NOT NULL,
-    PriceTotal INTEGER NOT NULL,
+    PriceTotal DECIMAL(4, 2) NOT NULL,
     FOREIGN KEY (OrderID) REFERENCES Orders(OrderID),
     FOREIGN KEY (ItemID) REFERENCES Items(ItemID)
 );
